# api-document.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API Document",
        "description": "Mota API Service",
        "license": {
            "name": "API Licence",
            "url": "http://domain.vn/license"
        },
        "version": "v1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "server test"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/user/create": {
            "post": {
                "tags": [
                    "User Controller"
                ],
                "summary": "Add user",
                "description": "Api create new user",
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/{id}/block": {
            "patch": {
                "tags": [
                    "User Controller"
                ],
                "summary": "Block user",
                "description": "Api block user",
                "operationId": "blockUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}/active": {
            "patch": {
                "tags": [
                    "User Controller"
                ],
                "summary": "Active user",
                "description": "Api active user",
                "operationId": "activeUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User Controller"
                ],
                "summary": "Get user",
                "description": "Api get user by Id",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/": {
            "get": {
                "tags": [
                    "User Controller"
                ],
                "summary": "getAllUser",
                "description": "Api get all user",
                "operationId": "getAllUsers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserDTO"
                                    }
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Get all user successfully",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "ex name": {
                                        "summary": "ex summary",
                                        "description": "ex name"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/recent/user/{id}": {
            "get": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "getMovieHistoryByUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WatchHistoryProjection"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/movies/user/{id}": {
            "get": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "getMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MovieProjection"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/movie/{id}/user/{userid}": {
            "get": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "getMovieByUserID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "userid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MovieDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/favorite/user/{id}": {
            "get": {
                "tags": [
                    "favorite-movie-controller"
                ],
                "operationId": "getFavoriteMoviesByUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MovieDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EpisodeDTO": {
                "type": "object",
                "properties": {
                    "episodeNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "videoUrl": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "MovieDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "country": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    },
                    "ageLimit": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "trailerUrl": {
                        "type": "string"
                    },
                    "posterUrl": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "single",
                            "series"
                        ]
                    },
                    "releaseDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "actors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "genres": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "episodes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EpisodeDTO"
                        }
                    },
                    "continueWatching": {
                        "$ref": "#/components/schemas/WatchHistoryDTO"
                    }
                }
            },
            "WatchHistoryDTO": {
                "type": "object",
                "properties": {
                    "episodeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "watchedSeconds": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "lastWatchedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "CreateUserRequest": {
                "required": [
                    "email",
                    "fullName",
                    "password",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "RegisterDTO": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    }
                }
            },
...
            "UserDTO": {
                "required": [
                    "email",
                    "fullName",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "blocked"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "fullName": {
                        "type": "string"
                    }
                }
            },
...
            "WatchHistoryProjection": {
                "type": "object",
                "properties": {
                    "episodeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "lastWatchedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "posterUrl": {
                        "type": "string"
                    },
                    "watchedSeconds": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "episodeNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "movieId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "MovieProjection": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "posterUrl": {
                        "type": "string"
                    },
                    "watchedSeconds": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "movieId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}